+++ packages/google-vertex/src/google-vertex-language-model.ts
@@ -12,6 +12,7 @@ import {
   GenerateContentResponse,
   GenerationConfig,
   Part,
+  SafetySetting,
   VertexAI,
 } from '@google-cloud/vertexai';
 import { convertToGoogleVertexContentRequest } from './convert-to-google-vertex-content-request';
@@ -99,6 +100,9 @@ export class GoogleVertexLanguageModel implements LanguageModelV1 {
             model: this.modelId,
             generationConfig,
             tools: prepareTools(mode),
+            safetySettings: this.settings.safetySettings as
+              | undefined
+              | Array<SafetySetting>,
           }),
           contentRequest: convertToGoogleVertexContentRequest(prompt),
           warnings,
@@ -316,6 +320,10 @@ function getToolCallsFromParts({
   parts: Part[];
   generateId: () => string;
 }) {
+  if (parts == null) {
+    return undefined; // parts are sometimes undefined when using safety settings
+  }
+
   return parts.flatMap(part =>
     part.functionCall == null
       ? []
@@ -328,7 +336,11 @@ function getToolCallsFromParts({
   );
 }
 
-function getTextFromParts(parts: Part[]) {
+function getTextFromParts(parts: Part[] | undefined) {
+  if (parts == null) {
+    return undefined; // parts are sometimes undefined when using safety settings
+  }
+
   const textParts = parts.filter(part => 'text' in part) as Array<
     Part & { text: string }
   >;