+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/google-vertex
 
+## 0.0.6
+
+### Patch Changes
+
+- 3a7a4ab6: fix (provider/vertex): fix undefined parts handling
+
+## 0.0.5
+
+### Patch Changes
+
+- 7cab5e9c: feat (provider/vertex): add safety setting option on models
+
 ## 0.0.4
 
 ### Patch Changes